Add logout functionality
